refactor: request dto 잘못된 타입 변경 #8

+++ be/src/main/java/buddyguard/mybuddyguard/weight/contoller/request/WeightCreateRequest.java
@@ -3,7 +3,7 @@ package buddyguard.mybuddyguard.weight.contoller.request;
 import java.time.LocalDateTime;
 
 public record WeightCreateRequest(
-        Integer petId,
+        Long petId,
         LocalDateTime recordedAt,
         Double weight,
         String description
